Arego libreria SweetAlert y try-catch

+++ js/checkout.js
@@ -1,4 +1,14 @@
-let carrito = JSON.parse(localStorage.getItem("carrito")) || []
+let carrito = []
+try {
+    const datos = localStorage.getItem("carrito")
+
+    if(datos != null)
+        carrito=JSON.parse(datos)
+} catch(error)
+{
+    console.error("Error al recuperar la informacion del carrito", error)
+    carrito=[]
+}
 
 const form = document.getElementById("form_checkout")
 const resumen = document.getElementById("resumen_compra")
@@ -16,7 +26,7 @@ form.onsubmit = function (evento)
     const direccion = document.getElementById("direccion").value
 
     const metodoPagoCredito = document.getElementById("credito")
-    const metodoPagoDebito = document.getElementById("Debito")
+    const metodoPagoDebito = document.getElementById("debito");
 
     let metodoPago = ""
     if(metodoPagoCredito.checked)
@@ -27,32 +37,55 @@ form.onsubmit = function (evento)
         
     if(carrito.length === 0)
     {
-        mensajeError.classList.add("mostrar");
-        mensajeError.textContent = "Your cart is empty"
+        Swal.fire({
+          title: "Your cart is empty!",
+          text: "Add products to continue.",
+          icon: "warning",
+        });
         return
     }
 
-    let total = 0
-    let listaProductos = ""
+    Swal.fire({
+      title: "Are you sure you want to complete the purchase?",
+      text: "Please confirm to finalize your order.",
+      icon: "question",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, confirm purchase",
+      cancelButtonText: "No, cancel"
+    }).then((result) => {
+      if (result.isConfirmed) 
+        {
+            let total = 0;
+            let listaProductos = "";
 
-    carrito.forEach(function (producto) {
-        const subtotal = producto.precio * producto.cantidad
-        total += subtotal
-        listaProductos += `<li>${producto.nombre} x${producto.cantidad} - $${subtotal}</li>` 
+            carrito.forEach(function (producto) {
+            const subtotal = producto.precio * producto.cantidad;
+            total += subtotal;
+            listaProductos += `<li>${producto.nombre} x${producto.cantidad} - $${subtotal}</li>`;
+            });
+
+            resumen.classList.add("mostrar");
+            resumen.innerHTML = `<h2> Purchase Summary</h2>
+                                <p>Name: ${nombre} ${apellido}</p>
+                                <p>Email: ${email}</p>
+                                <p>Adress: ${direccion}</p>
+                                <p>Payment Method: ${metodoPago}</p>
+                                <h3>Products:</h3>   
+                                <ul>${listaProductos}</ul>
+                                <h3>Total: $${total}</h3>
+                                <p id="texto-compra-final">Thank you for your purchase!<p>                 
+            `;
+            carrito = [];
+            localStorage.setItem("carrito", JSON.stringify(carrito));
+            form.reset();
+
+            Swal.fire({
+              title: "Purchase Confirmed",
+              text: "Your order has been successfully processed.",
+              icon: "success",
+            });
+        }
     });
-
-    resumen.classList.add("mostrar");
-    resumen.innerHTML = `<h2> Purchase Summary</h2>
-                         <p>Name: ${nombre} ${apellido}</p>
-                         <p>Email: ${email}</p>
-                         <p>Adress: ${direccion}</p>
-                         <p>Payment Method: ${metodoPago}</p>
-                         <h3>Products:</h3>   
-                         <ul>${listaProductos}</ul>
-                         <h3>Total: $${total}</h3>
-                         <p id="texto-compra-final">Thank you for your purchase!<p>                 
-    `
-    carrito = []
-    localStorage.setItem("carrito",JSON.stringify(carrito))
-    form.reset()
 }
